chore: refactor chatInputBox and fix useSocket bug
